+++ polaris-infrastructure/src/main/java/com/achobeta/infrastructure/dao/po/UserPO.java
@@ -39,7 +39,7 @@ public class UserPO {
     /**
      * 用户性别
      */
-    private Byte gender;
+    private Integer gender;
     /**
      * 用户身份证号
      */
